ui: System logs split view
