+++ core/ui/src/components/LogOutput.vue
@@ -2,6 +2,7 @@
   <div
     class="logs-output cv-code-snippet-multiline"
     :class="[
+      { 'reduced-output-height': numSearches > 1 && !verticalLayout },
       `bx--snippet`,
       `bx--snippet--multi`,
       {
@@ -22,7 +23,23 @@
       :paragraph="true"
       :line-count="4"
     ></cv-skeleton-text>
-    <div v-else class="bx--snippet-container" ref="logsContainer">
+    <NsEmptyState
+      v-else-if="noLogsFound"
+      :title="$t('system_logs.no_log_found')"
+      :animationData="GhostLottie"
+      animationTitle="ghost"
+      :loop="1"
+      class="margin-auto"
+    >
+      <template #description>
+        <div>{{ $t("system_logs.try_changing_search_filters") }}</div>
+      </template>
+    </NsEmptyState>
+    <div
+      v-else
+      class="bx--snippet-container"
+      :ref="'logsContainer-' + searchId"
+    >
       <pre><slot /></pre>
     </div>
   </div>
@@ -30,11 +47,12 @@
 
 <script>
 import { carbonPrefixMixin, themeMixin } from "@carbon/vue/src/mixins";
+import { UtilService, LottieService } from "@nethserver/ns8-ui-lib";
 
 export default {
   name: "LogOutput",
   props: {
-    id: {
+    searchId: {
       type: String,
       required: true,
     },
@@ -42,25 +60,28 @@ export default {
       type: Boolean,
       default: true,
     },
+    numSearches: {
+      type: Number,
+      default: 1,
+    },
+    verticalLayout: Boolean,
     loading: Boolean,
     wrapText: Boolean,
     light: Boolean,
     disabled: Boolean,
+    noLogsFound: Boolean,
   },
-  mixins: [carbonPrefixMixin, themeMixin],
+  mixins: [carbonPrefixMixin, themeMixin, UtilService, LottieService],
   watch: {
-    id: function () {
-      //// handle reqId
-      this.$root.$on("logsStart", this.logsUpdated);
+    searchId: function () {
+      this.$root.$on(`logsStart-${this.searchId}`, this.logsUpdated);
     },
   },
   created() {
-    //// handle reqId
-    this.$root.$on("logsStart", this.logsUpdated);
+    this.$root.$on(`logsStart-${this.searchId}`, this.logsUpdated);
   },
   beforeDestroy() {
-    //// handle reqId
-    this.$root.$off("logsStart");
+    this.$root.$off(`logsStart-${this.searchId}`);
   },
   methods: {
     // handleScroll: function (el) { ////
@@ -86,11 +107,14 @@ export default {
         if (this.scrollToBottom) {
           this.scrollToBottomOfContainer();
         }
-      }, 100);
+      }, 50);
     },
     scrollToBottomOfContainer() {
-      const el = this.$refs.logsContainer;
-      el.scrollTop = el.scrollHeight;
+      const el = this.$refs[`logsContainer-${this.searchId}`];
+
+      if (el) {
+        el.scrollTop = el.scrollHeight;
+      }
     },
   },
 };
@@ -98,6 +122,20 @@ export default {
 
 <style scoped lang="scss">
 @import "../styles/carbon-utils";
+
+.logs-output.bx--snippet--multi {
+  max-width: none;
+  min-height: 4rem;
+  max-height: 35rem;
+}
+
+.logs-output.reduced-output-height.bx--snippet--multi {
+  max-height: 20rem;
+}
+
+.margin-auto {
+  margin: auto;
+}
 </style>
 
 <style lang="scss">
@@ -107,5 +145,15 @@ export default {
 
 .bx--snippet--multi .bx--snippet-container pre {
   overflow-x: visible;
+}
+
+// remove fade effect on right border of code snippet
+.system-logs .logs-output.bx--snippet--multi .bx--snippet-container pre::after {
+  background-image: none;
+}
+
+// show scrollbar
+.system-logs .logs-output.bx--snippet--multi .bx--snippet-container {
+  overflow-y: auto !important;
 }
 </style>
