+++ core/ui/src/styles/_carbon-utils.scss
@@ -4,6 +4,7 @@
 @import "~carbon-components/scss/globals/scss/vars.scss";
 @import "./branding";
 
+$breakpoint-x-large: 1311px;
 $breakpoint-large: 1055px;
 $breakpoint-medium: 671px;
 $breakpoint-small: 319px;
